# server/swagger.yaml
openapi: 3.0.3
info:
  title: DLT TalentHub API
  description: |
    Backend API for DLT TalentHub - connecting talents with job opportunities in the DLT Africa ecosystem.
    
    ## Overview
    This API provides endpoints for user authentication, talent management, job postings, recruitment, and administrative functions.
    
    ## Authentication
    Most endpoints require authentication using JWT Bearer tokens. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    API requests are rate-limited to 100 requests per 15 minutes per IP address.
    
    ## Error Responses
    All error responses follow a consistent format:
    ```json
    {
      "status": "error",
      "message": "Error description",
      "errors": ["Detailed error messages"]
    }
    ```
  version: 1.0.0
  contact:
    name: DLT TalentHub Team
    email: support@dlt-talenthub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:5000
    description: Development server
  - url: https://api.dlt-talenthub.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /api/recruiters/jobs:
    get:
      summary: Get jobs posted by the recruiter
      description: Get all jobs posted by the current recruiter
      tags: [Recruiters]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, draft, paused, closed]
          description: Filter by job status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of recruiter's jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /api/jobs:
    post:
      summary: Create a new job posting
      description: Create a new job posting (recruiters and admins only)
      tags: [Jobs]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobCreateRequest'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/Job'
        '401':
          description: Not authenticated or not authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      summary: Get all jobs
      description: Get a paginated list of all active jobs
      tags: [Jobs]
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
        - in: query
          name: status
          schema:
            type: string
            enum: [active, draft, paused, closed]
          description: Filter by job status
        - in: query
          name: type
          schema:
            type: string
            enum: [full-time, part-time, contract, freelance, internship]
          description: Filter by job type
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /api/admin/users:
    get:
      summary: Get all users
      description: Get a paginated list of all users in the system
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [talent, recruiter, admin]
          description: Filter by user role
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
          description: Filter by user status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of all users
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      users:
                        type: array
                        items:
                          $ref: '#/components/schemas/User'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /api/admin/jobs:
    get:
      summary: Get all jobs
      description: Get a paginated list of all jobs in the system
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [active, draft, paused, closed]
          description: Filter by job status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of all jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      jobs:
                        type: array
                        items:
                          $ref: '#/components/schemas/Job'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
...
  /api/admin/notifications:
    get:
      summary: Get system notifications
      description: Get system-wide notifications and alerts
      tags: [Admin]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: Filter by notification type
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: System notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/notifications:
    get:
      summary: Get user notifications
      description: Get notifications for the current user
      tags: [Notifications]
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: type
          schema:
            type: string
          description: Filter by notification type
        - in: query
          name: isRead
          schema:
            type: boolean
          description: Filter by read status
        - in: query
          name: page
          schema:
            type: integer
            default: 1
            minimum: 1
          description: Page number for pagination
        - in: query
          name: limit
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: Number of results per page
      responses:
        '200':
          description: List of notifications
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: object
                    properties:
                      notifications:
                        type: array
                        items:
                          $ref: '#/components/schemas/Notification'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
        '401':
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        _id:
          type: string
          description: User ID
        firstName:
          type: string
          description: User's first name
        lastName:
          type: string
          description: User's last name
        email:
          type: string
          format: email
          description: User's email address
        role:
          type: string
          enum: [talent, recruiter, admin]
          description: User's role
        phone:
          type: string
          description: User's phone number
        bio:
          type: string
          description: User's bio
        location:
          type: string
          description: User's location
        avatar:
          type: string
          description: URL to user's avatar image
        isActive:
          type: boolean
          description: Whether the user account is active
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          description: Error message
        errors:
          type: array
          items:
            type: string
          description: Detailed error messages
...
    Job:
      type: object
      properties:
        _id:
          type: string
          description: Job ID
        title:
          type: string
          description: Job title
        description:
          type: string
          description: Job description
        company:
          type: object
          properties:
            name:
              type: string
              description: Company name
            logo:
              type: string
              description: Company logo URL
            website:
              type: string
              format: uri
              description: Company website
        type:
          type: string
          enum: [full-time, part-time, contract, freelance, internship]
          description: Job type
        category:
          type: string
          description: Job category
        skills:
          type: array
          items:
            type: string
          description: Required skills
        requirements:
          type: array
          items:
            type: string
          description: Job requirements
        location:
          type: object
          properties:
            city:
              type: string
              description: City
            country:
              type: string
              description: Country
            remote:
              type: boolean
              description: Whether remote work is allowed
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum salary
            max:
              type: number
              minimum: 0
              description: Maximum salary
            currency:
              type: string
              default: USD
              description: Salary currency
        status:
          type: string
          enum: [active, draft, paused, closed]
          description: Job status
        isUrgent:
          type: boolean
          description: Whether job is urgent
        isFeatured:
          type: boolean
          description: Whether job is featured
        recruiterId:
          type: string
          description: ID of the recruiter who posted the job
        views:
          type: integer
          minimum: 0
          description: Number of job views
        applications:
          type: integer
          minimum: 0
          description: Number of applications
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    JobCreateRequest:
      type: object
      required:
        - title
        - description
        - company
        - type
        - category
      properties:
        title:
          type: string
          minLength: 5
          maxLength: 200
          description: Job title
        description:
          type: string
          minLength: 50
          maxLength: 5000
          description: Job description
        company:
          type: object
          required:
            - name
          properties:
            name:
              type: string
              description: Company name
            logo:
              type: string
              format: uri
              description: Company logo URL
            website:
              type: string
              format: uri
              description: Company website
        type:
          type: string
          enum: [full-time, part-time, contract, freelance, internship]
          description: Job type
        category:
          type: string
          description: Job category
        skills:
          type: array
          items:
            type: string
          description: Required skills
        requirements:
          type: array
          items:
            type: string
          description: Job requirements
        location:
          type: object
          properties:
            city:
              type: string
              description: City
            country:
              type: string
              description: Country
            remote:
              type: boolean
              description: Whether remote work is allowed
        salary:
          type: object
          properties:
            min:
              type: number
              minimum: 0
              description: Minimum salary
            max:
              type: number
              minimum: 0
              description: Maximum salary
            currency:
              type: string
              default: USD
              description: Salary currency
        isUrgent:
          type: boolean
          description: Whether job is urgent
        isFeatured:
          type: boolean
          description: Whether job is featured
...
    Notification:
      type: object
      properties:
        _id:
          type: string
          description: Notification ID
        userId:
          type: string
          description: User ID
        type:
          type: string
          description: Notification type
        title:
          type: string
          description: Notification title
        message:
          type: string
          description: Notification message
        data:
          type: object
          description: Additional notification data
        isRead:
          type: boolean
          description: Whether notification is read
        createdAt:
          type: string
          format: date-time
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          description: Current page number
        limit:
          type: integer
          minimum: 1
          description: Number of items per page
        total:
          type: integer
          minimum: 0
          description: Total number of items
        pages:
          type: integer
          minimum: 0
          description: Total number of pages
        hasNext:
          type: boolean
          description: Whether there is a next page
        hasPrev:
          type: boolean
          description: Whether there is a previous page
...